+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Allow module config value on/off as boolean.
 - Show why a modules is installed in module info.
 - On app-graph, create SVG next to PNG and graph file.
+- Add command app-module-config-uncache.
 
 ## v1.0.2
 - Fix structure migration bug in module update.